hola_mundo/requirements.txt
dspy-ai
litellm>=1.10.1
google-generativeai>=0.3.0
flask>=2.0.0
python-dotenv>=0.19.0 